Fix XmlResultReporterTest
Currently, it's a fluke that this test passes — it just happens to be the case
that, though the temporary file that is created prevents a directory from being
created in its place, the test later uses the name of the file and everything
happens to work out.  This change fixes that by using the temporary file only
as a name, as was originally intended.

Change-Id: Id0e1e62be5bbeb225bde5be6348f0c8ad42969d6