Better handling for native test timeout.
Previously, a test run would abort if a native test execution timed out.
This is undesirable since a test timing out is an expected condition
(especially for stress tests).

This occurred because the underlying ITestDevice.executeShellCommand method
would always throw DeviceUnreponsiveException if command timed out. Change the
semantics of this method to allow caller to control this retry-and-throw
behavior.

Change-Id: I5c07f98831f67883ef10882250687c48d3ae790e