Add manual tests to Bluetooth stress tests.
These tests need some human intervention, such as running an
instrumentation on another device, putting a headset or keyboard into
pairing mode, etc.  These tests are therefore not run by default, they
print an instruction to stdout, and then require the user to follow the
instruction and press Enter before the test continues.  The tests are
then run, with results posted to sponge and the release dashboard like
the other Bluetooth stress tests.

Change-Id: I47ddf288102a9431c2da2f2a3c33592d988cb68b