Treat build ID as a String.
build/core/version_defaults.mk says that BUILD_NUMBER "can be an
arbitrary string (to allow for source control that uses something other
than numbers)". Since this is the make variable whose value ultimately
makes it to the target as the ro.build.version.incremental property in
devices, we'll treat is as a String.

Change-Id: I25a0e119b7dc344cfa12ee34b4ec857b6585c21b