Move up init of logger to before IBuildProvider.getBuild.
Previously, TF had this perplexing behavior where Log calls would
only be directed to the configuration's logger class until *after*
the IBuildProvider.getBuild stage. This made it difficult for users
to see log events that occurred during this stage, without fumbling
around with the global log level.

The reason this was done was to avoid saving a bunch of spurious logs
when IBuildProvider.getBuild was called but no build was available.

This commit satisfies both constraints by:
  - moving registration of the config logger to right at the start of
Invoker.invoke, so log messages during IBuildProvider.getBuild will
get directed properly
  - when IBuildProvider.getBuild fails (ie invocation is not saved),
dump the contents of the config'slogger to the global log, so its
contents are not lost.

Change-Id: Id840d33ccf169eadb66159d898afa03258eb5e83