add a new target preparer to push list of items from tests zip
added TestFilePushSetup to push a list of files or directories
from tests zip file onto device.

TestFilePushSetupFuncTest is a functional test.

As a testing fixture for TestFilePushSetupFuncTest, added
FakeTestsZipFolder to manage faking an expanded tests zip file
with contents backed by empty files. The contents of it are
populated based on a manifest provided to ctor.

Also added FakeTestsZipFolderTest...

Change-Id: I65f0c2d9cb33ca2dd7791aec7549b54d4641d03f