Quit after running initial command in non-interactive mode
This makes behavior consistent between the cases where:
1) Console is not available at all
2) Console is available, but reading from stdin fails

There's a question here of whether we want to support running TF as
a daemon in the background.  Currently, the answer is "no," as
there's currently no way to kill TF in a way that would allow it to
clean up temporary files and perform other cleanup.

Bug: 6363212
Change-Id: I122ccde51cac9aa8b619be7c1be7d13b8a04fe95