UI Automator test harness
This test harness includes an runner, and a new test type.

The runner takes care of invoking the on device command, parsing
the output

The test type takes care of passing down the parameters, and
common task such as taking a screen shot and bugreport when test
fails.

A sample config which drives demo test (source available in
Android tree) is also included.

Change-Id: I1e39c5965c630e28cd4874293790c89a2cd3e77e