Key procrank on PID, not app name.
An app or command line name of a process can have several lines, each
with unique PIDs.  Previously, only the last instance of the app would
be recorded.  Now, we key on PID which will be unique, and store the app
name as an attribute along with PSS, USS, RSS, and VSS.

Change-Id: Ibf1dfb3372a38609a978352188eaf36e51868b5e