Fix finding all apps to uninstall logic for certain devices.
Previously tradefed would assume all apps not installed on /system
are uninstallable. This is not true on some Android devices, which
has some apks on /vendor.

This commit changes the tf logic to use the 'dumpsys package p'
output to find uninstallable packages.

Note that neither 'pm list packages -3' or 'pm list packages -s' gives
the required info, because neither handles the case of 'a system app that
has had an update installed to /data'. Plus the fact that
'pm list packages -3' is only available on ICS and above.

Bug 8210531

Change-Id: I2eb6a25f16c4491ffc1dddc7d135a05ea92a777b